requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
python-dateutil==2.8.2
pytz==2023.3

# News and Data
feedparser==6.0.10
beautifulsoup4==4.12.2
newspaper3k==0.2.8

# Text-to-Speech
gTTS==2.3.2
pyttsx3==2.90

# Video and Audio Processing
moviepy==1.0.3
pydub==0.25.1

# Utilities
tqdm==4.66.1
python-crontab==3.0.0
python-slugify==8.0.1
loguru==0.7.0
